database changes

+++ db/migrate/20220415172634_create_users.rb
@@ -3,8 +3,8 @@ class CreateUsers < ActiveRecord::Migration[6.0]
     create_table :users do |t|
       t.string :name
       t.string :address
-      t.integer :phone, limit: 10
-      t.integer :adhar_number, limit: 16
+      t.bigint :phone, limit: 10
+      t.bigint :adhar_number, limit: 16
 
       t.timestamps
     end
